Resolve duplicate Trainer, MultipleDatasetTrainer, and Predictor classes
