+++ src/gnn_reco/components/utils.py
@@ -14,169 +14,6 @@ from gnn_reco.data.sqlite_dataset import SQLiteDataset
 from gnn_reco.models import Model
 
 # @TODO >>> RESOLVE DUPLICATION WRT. src/gnn_reco/models/training/{callbacks,trainers,utils}.py
-class MultipleDatasetsTrainer(object):
-    def __init__(self, training_loaders, validation_loaders, num_training_batches, num_validation_batches,optimizer, n_epochs, loss_func, target, device, scheduler = None, patience = 10, early_stopping = True):
-        self.validation_dataloaders = validation_loaders
-        self.training_dataloaders = training_loaders
-        self.num_training_batches = num_training_batches
-        self.num_validation_batches = num_validation_batches
-        if early_stopping:
-            self._early_stopping_method = EarlyStopping(patience = patience)
-        self.optimizer = optimizer
-        self.scheduler = scheduler
-        self.n_epochs = n_epochs
-        self.loss_func = loss_func
-        self.target = target
-        self.device = device
-
-    def __call__(self, model):
-        trained_model = self._train(model)
-        self._load_best_parameters(model)
-        return trained_model
-
-    def _train(self,model):
-        training_batches = self.num_training_batches
-        for epoch in range(self.n_epochs):
-            acc_loss = torch.tensor([0],dtype = float).to(self.device)
-            iteration = 1
-            model.train()
-            pbar = tqdm(total = training_batches, unit= 'batches') 
-            for training_dataloader in self.training_dataloaders:  
-                for batch_of_graphs in training_dataloader:
-                    batch_of_graphs.to(self.device)
-                    with torch.enable_grad():                                                                                                     
-                            self.optimizer.zero_grad()                                                   
-                            out                             = model(batch_of_graphs)   
-                            loss                            = self.loss_func(out, batch_of_graphs, self.target)                      
-                            loss.backward()                                                         
-                            self.optimizer.step()
-                    if self.scheduler != None:    
-                        self.optimizer.param_groups[0]['lr'] = self.scheduler.get_next_lr().item()
-                    acc_loss += loss
-                    iteration +=1
-                    pbar.update(1)
-                    pbar.set_description('epoch: %s || loss: %s'%(epoch, acc_loss.item()/iteration))
-            validation_loss = self._validate(model)
-            pbar.set_description('epoch: %s || loss: %s || valid loss : %s'%(epoch,acc_loss.item()/iteration, validation_loss.item()))
-            if self._early_stopping_method.step(validation_loss,model):
-                print('EARLY STOPPING: %s'%epoch)
-                break
-        return model
-        
-    def _validate(self,model):
-        acc_loss = torch.tensor([0],dtype = float).to(self.device)
-        model.eval()
-        iterations = 1
-        pbar_valid = tqdm(total = self.num_validation_batches, unit= 'batches')
-        pbar_valid.set_description('Validating..')
-        for validation_dataloader in self.validation_dataloaders:
-            for batch_of_graphs in validation_dataloader:
-                batch_of_graphs.to(self.device)
-                with torch.no_grad():                                                                                        
-                    out                             = model(batch_of_graphs)   
-                    loss                            = self.loss_func(out, batch_of_graphs, self.target)                             
-                    acc_loss += loss
-                iterations +=1
-                pbar_valid.update(1)
-        return acc_loss/iterations
-    
-    def _load_best_parameters(self,model):
-        return model.load_state_dict(self._early_stopping_method.GetBestParams())
-
-class Trainer(object):
-    def __init__(self, training_dataloader, validation_dataloader, optimizer, n_epochs, loss_func, target, device, scheduler = None, patience = 10, early_stopping = True):
-        self.training_dataloader = training_dataloader
-        self.validation_dataloader = validation_dataloader
-        if early_stopping:
-            self._early_stopping_method = EarlyStopping(patience = patience)
-        self.optimizer = optimizer
-        self.scheduler = scheduler
-        self.n_epochs = n_epochs
-        self.loss_func = loss_func
-        self.target = target
-        self.device = device
-
-    def __call__(self, model):
-        trained_model = self._train(model)
-        self._load_best_parameters(model)
-        return trained_model
-
-    def _train(self,model):
-        for epoch in range(self.n_epochs):
-            acc_loss = torch.tensor([0],dtype = float).to(self.device)
-            iteration = 1
-            model.train() 
-            pbar = tqdm(self.training_dataloader, unit= 'batches')
-            for batch_of_graphs in pbar:
-                batch_of_graphs.to(self.device)
-                with torch.enable_grad():                                                                                                     
-                        self.optimizer.zero_grad()                                                   
-                        out                             = model(batch_of_graphs)   
-                        loss                            = self.loss_func(out, batch_of_graphs, self.target)                      
-                        loss.backward()                                                         
-                        self.optimizer.step()
-                if self.scheduler != None:    
-                    self.optimizer.param_groups[0]['lr'] = self.scheduler.get_next_lr().item()
-                acc_loss += loss
-                if iteration == (len(pbar)):    
-                    validation_loss = self._validate(model)
-                    pbar.set_description('epoch: %s || loss: %s || valid loss : %s'%(epoch,acc_loss.item()/iteration, validation_loss.item()))
-                else:
-                    pbar.set_description('epoch: %s || loss: %s'%(epoch, acc_loss.item()/iteration))
-                iteration +=1
-            if self._early_stopping_method.step(validation_loss,model):
-                print('EARLY STOPPING: %s'%epoch)
-                break
-        return model
-            
-    def _validate(self,model):
-        print('validating')
-        acc_loss = torch.tensor([0],dtype = float).to(self.device)
-        model.eval()
-        for batch_of_graphs in self.validation_dataloader:
-            batch_of_graphs.to(self.device)
-            with torch.no_grad():                                                                                        
-                out                             = model(batch_of_graphs)   
-                loss                            = self.loss_func(out, batch_of_graphs, self.target)                             
-                acc_loss += loss
-        return acc_loss/len(self.validation_dataloader)
-    
-    def _load_best_parameters(self,model):
-        return model.load_state_dict(self._early_stopping_method.GetBestParams())
-         
-
-class Predictor(object):
-    def __init__(self, dataloader, target, device, output_column_names, post_processing_method = None):
-        self.dataloader = dataloader
-        self.target = target
-        self.output_column_names = output_column_names
-        self.device = device
-        self.post_processing_method = post_processing_method
-    def __call__(self, model):
-        self.model = model.to(self.device)
-        self.model.eval()
-        self.model.predict = True
-        if self.post_processing_method == None:
-            return self._predict()
-        else:
-            return self.post_processing_method(self._predict(),self.target)
-
-    def _predict(self):
-        predictions = []
-        event_nos   = []
-        target      = []
-        with torch.no_grad():
-            for batch_of_graphs in tqdm(self.dataloader, unit = 'batches'):
-                batch_of_graphs.to(self.device)
-                target.extend(batch_of_graphs[self.target].detach().cpu().numpy())
-                predictions.extend(self.model(batch_of_graphs).detach().cpu().numpy())
-                event_nos.extend(batch_of_graphs['event_no'].detach().cpu().numpy())
-        out = pd.DataFrame(data = predictions, columns = self.output_column_names)
-        out['event_no'] = event_nos
-        out[self.target] = target
-        return out
-     
-
 def make_train_validation_dataloader(db, selection, pulsemap, batch_size, FEATURES, TRUTH, num_workers, database_indices = None, persistent_workers = True, seed = 42):
     rng = np.random.RandomState(seed=seed)
     if isinstance(db, list):
